Отладочные функции перехали в единую библиотеку
