+++ Firmware/Source/Controller/Controller.c
@@ -2,7 +2,6 @@
 #include "Controller.h"
 //
 // Includes
-#include "DebugActions.h"
 #include "Board.h"
 #include "Delay.h"
 #include "DeviceProfile.h"
