--- Firmware/Source/Controller/DebugActions.h
@@ -1,12 +1,0 @@
-﻿#ifndef __DEBUGACTIONS_H
-#define __DEBUGACTIONS_H
-
-// Include
-#include "stdinc.h"
-
-void DBGACT_GenerateImpulseLineSync1();
-void DBGACT_GenerateImpulseLineSync2();
-void DBGACT_SetStatePOWRelay(bool NewState);
-void DBGACT_SetStateCTRLRelay(bool NewState);
-
-#endif //__DEBUGACTIONS_H
