+++ Firmware/Source/Controller/Diagnostic.c
@@ -7,7 +7,12 @@
 #include "LowLevel.h"
 #include "Delay.h"
 #include "Controller.h"
-#include "DebugActions.h"
+
+// Forward functions
+void DBGACT_GenerateImpulseLineSync1();
+void DBGACT_GenerateImpulseLineSync2();
+void DBGACT_SetStatePOWRelay(bool NewState);
+void DBGACT_SetStateCTRLRelay(bool NewState);
 
 // Functions
 bool DIAG_HandleDiagnosticAction(uint16_t ActionID, uint16_t *pUserError)
@@ -43,3 +48,31 @@ bool DIAG_HandleDiagnosticAction(uint16_t ActionID, uint16_t *pUserError)
 	}
 	return true;
 }
+//-----------------------------
+
+void DBGACT_GenerateImpulseLineSync1()
+{
+	LL_SetSync1State(true);
+	DELAY_MS(10);
+	LL_SetSync1State(false);
+}
+//-----------------------------
+
+void DBGACT_GenerateImpulseLineSync2()
+{
+	LL_SetSync2State(true);
+	DELAY_MS(10);
+	LL_SetSync2State(false);
+}
+//-----------------------------
+
+void DBGACT_SetStatePOWRelay(bool NewState)
+{
+	LL_ConnectPOWRelay(NewState);
+}
+//-----------------------------
+void DBGACT_SetStateCTRLRelay(bool NewState)
+{
+	LL_ConnectCTRLRelay(NewState);
+}
+//-----------------------------
